+++ modules/apps/asset/asset-categories-navigation-web/bnd.bnd
@@ -4,6 +4,6 @@ Bundle-Version: 1.0.0
 Include-Resource:\
 	classes,\
 	META-INF/resources=src/META-INF/resources
-Web-ContextPath: /asset-categories-navigation-web
 Require-Capability:\
 	osgi.extender;osgi.extender="jsp.taglib";uri="http://liferay.com/tld/ddm";version:Version="1.0.0"
+Web-ContextPath: /asset-categories-navigation-web
