+++ modules/apps/blogs/blogs-web/bnd.bnd
@@ -9,6 +9,6 @@ Export-Package:\
 Include-Resource:\
 	classes,\
 	META-INF/resources=src/META-INF/resources
-Web-ContextPath: /blogs-web
 Require-Capability:\
 	osgi.extender;osgi.extender="jsp.taglib";uri="http://liferay.com/tld/ddm";version:Version="1.0.0"
+Web-ContextPath: /blogs-web
