+++ modules/apps/dynamic-data-lists/dynamic-data-lists-web/bnd.bnd
@@ -8,6 +8,6 @@ Export-Package:\
 Include-Resource:\
 	classes,\
 	META-INF/resources=src/META-INF/resources
-Web-ContextPath: /ddl-web
 Require-Capability:\
 	osgi.extender;osgi.extender="jsp.taglib";uri="http://liferay.com/tld/ddm";version:Version="1.0.0"
+Web-ContextPath: /ddl-web
