+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-taglib/src/META-INF/liferay-ddm.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0"?>
 
-<taglibs short-name="ddm" uri="http://liferay.com/tld/ddm" tlib-version="1.0">
+<taglibs short-name="liferay-ddm" uri="http://liferay.com/tld/ddm" tlib-version="1.0">
 	<authors>
 		<author>Bruno Basto</author>
 	</authors>
