+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-taglib/src/META-INF/resources/liferay-ddm.tld
@@ -2,11 +2,11 @@
 
 <taglib xmlns="http://java.sun.com/xml/ns/j2ee" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" version="2.0" xsi:schemaLocation="http://java.sun.com/xml/ns/j2ee http://java.sun.com/xml/ns/j2ee/web-jsptaglibrary_2_0.xsd">
 	<tlib-version>1.0</tlib-version>
-	<short-name>ddm</short-name>
+	<short-name>liferay-ddm</short-name>
 	<uri>http://liferay.com/tld/ddm</uri>
 	<tag>
 		<name>html</name>
-		<tag-class>com.liferay.dynamic.data.mapping.taglib.ui.HTMLTag</tag-class>
+		<tag-class>com.liferay.dynamic.data.mapping.taglib.HTMLTag</tag-class>
 		<body-content>JSP</body-content>
 		<attribute>
 			<name>checkRequired</name>
@@ -72,7 +72,7 @@
 	</tag>
 	<tag>
 		<name>html-field</name>
-		<tag-class>com.liferay.dynamic.data.mapping.taglib.ui.HTMLFieldTag</tag-class>
+		<tag-class>com.liferay.dynamic.data.mapping.taglib.HTMLFieldTag</tag-class>
 		<body-content>JSP</body-content>
 		<attribute>
 			<name>classNameId</name>
@@ -132,7 +132,7 @@
 	</tag>
 	<tag>
 		<name>template-renderer</name>
-		<tag-class>com.liferay.dynamic.data.mapping.taglib.ui.TemplateRendererTag</tag-class>
+		<tag-class>com.liferay.dynamic.data.mapping.taglib.TemplateRendererTag</tag-class>
 		<body-content>JSP</body-content>
 		<attribute>
 			<name>className</name>
@@ -174,7 +174,7 @@
 	</tag>
 	<tag>
 		<name>template-selector</name>
-		<tag-class>com.liferay.dynamic.data.mapping.taglib.ui.TemplateSelectorTag</tag-class>
+		<tag-class>com.liferay.dynamic.data.mapping.taglib.TemplateSelectorTag</tag-class>
 		<body-content>JSP</body-content>
 		<attribute>
 			<name>className</name>
