+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-taglib/src/META-INF/resources/taglib/html_field/end.jsp
@@ -14,6 +14,6 @@
  */
 --%>
 
-<%@ include file="/taglib/ui/html_field/init.jsp" %>
+<%@ include file="/taglib/html_field/init.jsp" %>
 
 </div>
