+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-taglib/src/META-INF/resources/taglib/template_selector/page.jsp
@@ -14,11 +14,11 @@
  */
 --%>
 
-<%@ include file="/taglib/ui/template_selector/init.jsp" %>
+<%@ include file="/taglib/template_selector/init.jsp" %>
 
 <%
-long classNameId = GetterUtil.getLong((String)request.getAttribute("ddm:template-selector:classNameId"));
-DDMTemplate portletDisplayDDMTemplate = (DDMTemplate)request.getAttribute("ddm:template-selector:portletDisplayDDMTemplate");
+long classNameId = GetterUtil.getLong((String)request.getAttribute("liferay-ddm:template-selector:classNameId"));
+DDMTemplate portletDisplayDDMTemplate = (DDMTemplate)request.getAttribute("liferay-ddm:template-selector:portletDisplayDDMTemplate");
 
 long ddmTemplateGroupId = PortletDisplayTemplateManagerUtil.getDDMTemplateGroupId(themeDisplay.getScopeGroupId());
