+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-taglib/src/com/liferay/dynamic/data/mapping/taglib/TemplateRendererTag.java
@@ -12,9 +12,9 @@
  * details.
  */
 
-package com.liferay.dynamic.data.mapping.taglib.ui;
+package com.liferay.dynamic.data.mapping.taglib;
 
-import com.liferay.dynamic.data.mapping.taglib.ui.base.BaseTemplateRendererTag;
+import com.liferay.dynamic.data.mapping.taglib.base.BaseTemplateRendererTag;
 import com.liferay.portal.kernel.portletdisplaytemplate.PortletDisplayTemplateManagerUtil;
 import com.liferay.portal.kernel.util.WebKeys;
 import com.liferay.portal.theme.ThemeDisplay;
