+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-taglib/src/com/liferay/dynamic/data/mapping/taglib/base/BaseTemplateRendererTag.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.dynamic.data.mapping.taglib.ui.base;
+package com.liferay.dynamic.data.mapping.taglib.base;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.jsp.JspException;
@@ -128,13 +128,13 @@ public abstract class BaseTemplateRendererTag extends com.liferay.taglib.util.In
 		setNamespacedAttribute(request, "servletContext", _servletContext);
 	}
 
-	protected static final String _ATTRIBUTE_NAMESPACE = "ddm:template-renderer:";
+	protected static final String _ATTRIBUTE_NAMESPACE = "liferay-ddm:template-renderer:";
 
 	private static final String _END_PAGE =
-		"/taglib/ui/template_renderer/end.jsp";
+		"/taglib/template_renderer/end.jsp";
 
 	private static final String _START_PAGE =
-		"/taglib/ui/template_renderer/start.jsp";
+		"/taglib/template_renderer/start.jsp";
 
 	private java.lang.String _className = null;
 	private java.util.Map<java.lang.String, java.lang.Object> _contextObjects = null;
