+++ modules/apps/dynamic-data-mapping/dynamic-data-mapping-taglib/src/com/liferay/dynamic/data/mapping/taglib/base/BaseTemplateSelectorTag.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.dynamic.data.mapping.taglib.ui.base;
+package com.liferay.dynamic.data.mapping.taglib.base;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.jsp.JspException;
@@ -171,10 +171,10 @@ public abstract class BaseTemplateSelectorTag extends com.liferay.taglib.util.In
 		setNamespacedAttribute(request, "servletContext", _servletContext);
 	}
 
-	protected static final String _ATTRIBUTE_NAMESPACE = "ddm:template-selector:";
+	protected static final String _ATTRIBUTE_NAMESPACE = "liferay-ddm:template-selector:";
 
 	private static final String _PAGE =
-		"/taglib/ui/template_selector/page.jsp";
+		"/taglib/template_selector/page.jsp";
 
 	private java.lang.String _className = null;
 	private java.lang.String _defaultDisplayStyle = null;
