+++ modules/apps/site-navigation/site-navigation-breadcrumb-web/bnd.bnd
@@ -5,6 +5,6 @@ Export-Package: com.liferay.site.navigation.breadcrumb.web.constants
 Include-Resource:\
 	classes,\
 	META-INF/resources=src/META-INF/resources
-Web-ContextPath: /site-navigation-breadcrumb-web
 Require-Capability:\
 	osgi.extender;osgi.extender="jsp.taglib";uri="http://liferay.com/tld/ddm";version:Version="1.0.0"
+Web-ContextPath: /site-navigation-breadcrumb-web
